[Refactor] Split scan morsel interfaces
