+++ be/src/connector/data_source.h
@@ -21,7 +21,8 @@
 
 #include "common/runtime_profile.h"
 #include "common/statusor.h"
-#include "exec/pipeline/scan/morsel.h"
+#include "exec/pipeline/scan/morsel_fwd.h"
+#include "exec/pipeline/scan/scan_split_context.h"
 #include "exec/runtime_filter/runtime_filter_probe.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gen_cpp/PlanNodes_types.h"
