+++ be/src/exec/connector_scan_node.h
@@ -17,6 +17,7 @@
 #include <atomic>
 #include <memory>
 
+#include "base/concurrency/blocking_queue.hpp"
 #include "column/vectorized_fwd.h"
 #include "common/statusor.h"
 #include "connector/connector.h"
