+++ be/src/exec/pipeline/capture_version_operator.h
@@ -16,6 +16,7 @@
 
 #include "common/runtime_profile.h"
 #include "exec/pipeline/source_operator.h"
+#include "storage/rowset/rowset.h"
 
 namespace starrocks::pipeline {
