+++ be/src/exec/pipeline/scan/olap_meta_scan_prepare_operator.cpp
@@ -18,7 +18,9 @@
 
 #include "exec/pipeline/scan/meta_scan_context.h"
 #include "exec/pipeline/scan/meta_scan_operator.h"
+#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "gen_cpp/Types_types.h"
+#include "gutil/casts.h"
 #include "storage/olap_common.h"
 
 namespace starrocks::pipeline {
@@ -31,7 +33,8 @@ OlapMetaScanPrepareOperator::OlapMetaScanPrepareOperator(OperatorFactory* factor
           _scan_node(scan_node) {}
 
 Status OlapMetaScanPrepareOperator::_prepare_scan_context(RuntimeState* state) {
-    auto meta_scan_ranges = _morsel_queue->prepare_olap_scan_ranges();
+    auto* olap_morsel_queue = down_cast<OlapMorselQueue*>(_morsel_queue);
+    auto meta_scan_ranges = olap_morsel_queue->prepare_olap_scan_ranges();
     for (auto& scan_range : meta_scan_ranges) {
         MetaScannerParams params;
         params.scan_range = scan_range;
