+++ be/src/exec/pipeline/scan/scan_operator.h
@@ -22,6 +22,7 @@
 #include "exec/pipeline/topn_runtime_filter_back_pressure.h"
 #include "exec/query_cache/cache_operator.h"
 #include "exec/query_cache/lane_arbiter.h"
+#include "exec/query_cache/ticket_checker.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "exprs/chunk_predicate_evaluator.h"
