+++ be/src/storage/dictionary_cache_manager.cpp
@@ -14,6 +14,7 @@
 
 #include "storage/dictionary_cache_manager.h"
 
+#include "base/string/faststring.h"
 #include "exec/tablet_info.h"
 
 namespace starrocks {
