+++ be/src/storage/lake/tablet_reader.cpp
@@ -25,6 +25,10 @@
 #include "common/config_lake_fwd.h"
 #include "common/config_scan_io_fwd.h"
 #include "common/status.h"
+#include "exec/pipeline/scan/lake_split_context.h"
+#include "exec/pipeline/scan/scan_morsel.h"
+#include "exec/pipeline/scan/split_morsel_queue.h"
+#include "exec/pipeline/scan/split_scan_morsel.h"
 #include "gutil/stl_util.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
