+++ be/test/exec/pipeline/olap_scan_operator_test.cpp
@@ -15,6 +15,7 @@
 #include "exec/pipeline/scan/olap_scan_operator.h"
 
 #include "exec/olap_scan_node.h"
+#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "exec/pipeline/scan/olap_scan_prepare_operator.h"
 #include "gtest/gtest.h"
 #include "runtime/descriptors.h"
